+++ test/unit/data_frame_test.exs
@@ -218,7 +218,7 @@ defmodule SparkEx.DataFrameTest do
       {:ok, session} = SchemaSession.start_link(schema)
       df = DataFrame.new(session, {:sql, "SELECT 1", nil})
 
-      assert {:ok, [{"id", "BIGINT"}]} = DataFrame.dtypes(df)
+      assert {:ok, [{"id", "bigint"}]} = DataFrame.dtypes(df)
     end
 
     test "accepts {:ok, dataframe} for create_dataframe-style pipelines" do
@@ -235,7 +235,7 @@ defmodule SparkEx.DataFrameTest do
       {:ok, session} = SchemaSession.start_link(schema)
       df = DataFrame.new(session, {:sql, "SELECT 1", nil})
 
-      assert {:ok, [{"id", "BIGINT"}]} = DataFrame.dtypes({:ok, df})
+      assert {:ok, [{"id", "bigint"}]} = DataFrame.dtypes({:ok, df})
     end
   end